Created modules for interfaces and implementation; added spring flavor
