+++ cost/src/main/java/com/nulogy/nupack/object/ElectronicsLabour.java
@@ -1,0 +1,21 @@
+package com.nulogy.nupack.object;
+
+import java.math.BigDecimal;
+
+public class ElectronicsLabour extends BaseLabour {
+
+    private static final long serialVersionUID = 1L;
+    
+    public ElectronicsLabour() {
+        super();
+    }
+    
+    public ElectronicsLabour(int people) {
+        super(people);
+    }
+
+    @Override
+    public BigDecimal getMarkUp() {
+        return BigDecimal.valueOf(0.02).add(super.getMarkUp());
+    }
+}
