+++ cost/src/main/java/com/nulogy/nupack/object/PharmaceuticalLabour.java
@@ -1,0 +1,22 @@
+package com.nulogy.nupack.object;
+
+import java.math.BigDecimal;
+
+public class PharmaceuticalLabour extends BaseLabour {
+    
+    private static final long serialVersionUID = 1L;
+
+    public PharmaceuticalLabour() {
+        super();
+    }
+    
+    public PharmaceuticalLabour(int people) {
+        super(people);
+    }
+    
+    @Override
+    public BigDecimal getMarkUp() {
+        return BigDecimal.valueOf(0.075).add(super.getMarkUp());
+    }
+
+}
