+++ cost/src/main/java/com/nulogy/nupack/object/ValidationException.java
@@ -1,0 +1,11 @@
+package com.nulogy.nupack.object;
+
+public class ValidationException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+    
+    public ValidationException(String message) {
+        super(message);
+    }
+
+}
