+++ cost/src/main/java/com/nulogy/nupack/service/CostEstimationService.java
@@ -1,0 +1,22 @@
+package com.nulogy.nupack.service;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
+
+import com.nulogy.nupack.object.Labour;
+
+/**
+ * Service class for estimating labour cost
+ * @author George
+ *
+ */
+public interface CostEstimationService extends Serializable {
+
+    /**
+     * Estimate the final cost of labour 
+     * @param initalCost the initial cost
+     * @param labour the required labour
+     * @return the final cost 
+     */
+    BigDecimal estimate(BigDecimal initialCost, Labour labour);
+}
