--- src/main/java/com/nulogy/nupack/ElectronicsLabour.java
@@ -1,21 +1,0 @@
-package com.nulogy.nupack;
-
-import java.math.BigDecimal;
-
-public class ElectronicsLabour extends BaseLabour {
-
-    private static final long serialVersionUID = 1L;
-    
-    public ElectronicsLabour() {
-        super();
-    }
-    
-    public ElectronicsLabour(int people) {
-        super(people);
-    }
-
-    @Override
-    public BigDecimal getMarkUp() {
-        return BigDecimal.valueOf(0.02).add(super.getMarkUp());
-    }
-}
