--- src/main/java/com/nulogy/nupack/ValidationException.java
@@ -1,11 +1,0 @@
-package com.nulogy.nupack;
-
-public class ValidationException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-    
-    public ValidationException(String message) {
-        super(message);
-    }
-
-}
